EDITOR-#203 [FullStack] implement Searchbar

+++ frontend/src/components/AnnouncementList.tsx
@@ -8,6 +8,17 @@ import LoaderSpinner from "./LoaderSpinner";
 import { Button } from "@/components/ui/button";
 import { useUser } from "@/context/UserContext";
 import { useToast } from "./ui/use-toast";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 function AnnouncementList() {
   const navigate = useNavigate();
@@ -26,24 +37,22 @@ function AnnouncementList() {
 
   const announcements = data?.AllAnnouncements || [];
 
-  const handleDelete = (id: number) => {
-    if (window.confirm("Are you sure you want to delete this announcement?")) {
-      deleteAnnouncement({
-        variables: {
-          deleteAnnouncementId: id,
-        },
-      });
-      if (deleteLoading) return <LoaderSpinner />;
-      if (deleteError) {
-        toast({ title: `${deleteError.message}`, variant: "destructive" });
-      } else {
-        toast({ title: "Announcement deleted successfully!" });
-      }
+  const handleDelete = async (id: number) => {
+    await deleteAnnouncement({
+      variables: {
+        deleteAnnouncementId: id,
+      },
+    });
+    if (deleteLoading) return <LoaderSpinner />;
+    if (deleteError) {
+      toast({ title: `${deleteError.message}`, variant: "destructive" });
+    } else {
+      toast({ title: "Announcement deleted successfully!" });
     }
   };
   return (
     <>
-      {announcements.length && (
+      {announcements.length !== 0 ? (
         <div className="flex flex-col mx-3 border-2 border-blue-600 rounded-xl max-h-[500px] overflow-y-auto">
           {announcements.map(
             (announcement) =>
@@ -63,12 +72,35 @@ function AnnouncementList() {
                   </p>
                   {user?.isAdmin && (
                     <div className="flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2">
-                      <Button
-                        onClick={() => handleDelete(announcement.id)}
-                        className="text-red-400 border border-red-400 transform active:scale-90 transition-transform duration-200"
-                      >
-                        delete
-                      </Button>
+                      <AlertDialog>
+                        <AlertDialogTrigger asChild>
+                          <Button className="text-red-400 border border-red-400 transform active:scale-90 transition-transform duration-200">
+                            delete
+                          </Button>
+                        </AlertDialogTrigger>
+                        <AlertDialogContent className="text-white bg-black">
+                          <AlertDialogHeader>
+                            <AlertDialogTitle>
+                              Are you absolutely sure?
+                            </AlertDialogTitle>
+                            <AlertDialogDescription>
+                              This action cannot be undone. This will
+                              permanently delete the announcement.
+                            </AlertDialogDescription>
+                          </AlertDialogHeader>
+                          <AlertDialogFooter>
+                            <AlertDialogCancel className="text-sky-300 border border-sky-300 transform active:scale-90 transition-transform duration-200 bg-transparent hover:bg-primary/90 hover:text-sky-300">
+                              cancel
+                            </AlertDialogCancel>
+                            <AlertDialogAction
+                              className="text-red-400 border border-red-400 transform active:scale-90 transition-transform duration-200 bg-transparent hover:bg-primary/90"
+                              onClick={() => handleDelete(announcement.id)}
+                            >
+                              continue
+                            </AlertDialogAction>
+                          </AlertDialogFooter>
+                        </AlertDialogContent>
+                      </AlertDialog>
                       <Button
                         onClick={() =>
                           navigate(`/EditAnnouncement/${announcement.id}`)
@@ -83,6 +115,14 @@ function AnnouncementList() {
               )
           )}
         </div>
+      ) : (
+        <section className="flex flex-col size-full flex-center">
+          <div className="flex-center w-full flex-col mt-36">
+            <p className="w-full text-5xl text-center font-bold text-white p-2">
+              No Announcement
+            </p>
+          </div>
+        </section>
       )}
     </>
   );
